requirements.txt
playwright
python-dotenv
openai
gologin
asyncpg